# [performance] Optimize local timeline + local status count queries