# Description

> If this is a code change, please include a summary of what you've coded, and link to the issue(s) it closes/implements.
>
> If this is a documentation change, please briefly describe what you've changed and why.

Queries for counting local statuses and for fetching the local timeline have been annoyingly slow lately so I did some optimization. Count and timeline queries can now both use the updated `statuses_local_idx`, which should make things a lot faster.

Tested on both SQLite and Postgres. Postgres doesn't seem to use `statuses_local_idx` for local timeline queries in my testing, as it prefers to use `statuses_public_timeline_idx` and then filter local, but SQLite does, and both drivers use `statuses_local_idx` for the count query.

SQLite:

```
sqlite> CREATE INDEX IF NOT EXISTS "statuses_local_idx" ON "statuses" ("local", "visibility", "pending_approval", "id" DESC);
sqlite> PRAGMA analysis_limit=0; ANALYZE;
0
sqlite> .expert
sqlite> SELECT * FROM "statuses_local_count_view";
(no new indexes)

CO-ROUTINE statuses_local_count_view
SEARCH statuses USING COVERING INDEX statuses_local_idx (local=? AND visibility=? AND pending_approval=?)
SCAN statuses_local_count_view

sqlite> .expert
sqlite> SELECT "status"."id" FROM "statuses" AS "status" WHERE ("status"."local" = TRUE) AND ("status"."visibility" = 2) AND ("status"."pending_approval" = FALSE) AND ("status"."boost_of_id" IS NULL) AND ("status"."id" < '01JP2FR3VR7P63CZSSERJN7ZXJ') ORDER BY "status"."id" DESC LIMIT 20;
(no new indexes)

SEARCH status USING INDEX statuses_local_idx (local=? AND visibility=? AND pending_approval=? AND id<?)
```

Postgres:

```
postgres=# EXPLAIN ANALYZE SELECT * FROM "statuses_local_count_view";
                                                               QUERY PLAN                                                                
-----------------------------------------------------------------------------------------------------------------------------------------
 Aggregate  (cost=8.16..8.17 rows=1 width=8) (actual time=0.042..0.042 rows=1 loops=1)
   ->  Index Only Scan using statuses_local_idx on statuses  (cost=0.14..8.16 rows=1 width=0) (actual time=0.031..0.039 rows=21 loops=1)
         Index Cond: ((local = true) AND (visibility = ANY ('{2,3,4,5}'::integer[])) AND (pending_approval = false))
         Heap Fetches: 21
 Planning Time: 0.732 ms
 Execution Time: 0.065 ms
(6 rows)

postgres=# EXPLAIN ANALYZE SELECT "status"."id" FROM "statuses" AS "status" WHERE ("status"."local" = TRUE) AND ("status"."visibility" = 2) AND ("status"."pending_approval" = FALSE) AND ("status"."boost_of_id" IS NULL) AND ("status"."id" < '01JP2FR3VR7P63CZSSERJN7ZXJ') ORDER BY "status"."id" DESC LIMIT 20;
                                                                      QUERY PLAN                                                                       
-------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.14..8.16 rows=1 width=108) (actual time=0.026..0.040 rows=15 loops=1)
   ->  Index Scan using statuses_public_timeline_idx on statuses status  (cost=0.14..8.16 rows=1 width=108) (actual time=0.025..0.037 rows=15 loops=1)
         Index Cond: ((visibility = 2) AND (id < '01JP2FR3VR7P63CZSSERJN7ZXJ'::bpchar))
         Filter: (local AND (NOT pending_approval) AND (boost_of_id IS NULL))
         Rows Removed by Filter: 4
 Planning Time: 0.218 ms
 Execution Time: 0.053 ms
(7 rows)
```

## Checklist

Please put an x inside each checkbox to indicate that you've read and followed it: `[ ]` -> `[x]`

If this is a documentation change, only the first checkbox must be filled (you can delete the others if you want).

- [x] I/we have read the [GoToSocial contribution guidelines](https://github.com/superseriousbusiness/gotosocial/blob/main/CONTRIBUTING.md).
- [x] I/we have discussed the proposed changes already, either in an issue on the repository, or in the Matrix chat.
- [x] I/we have not leveraged AI to create the proposed changes.
- [x] I/we have performed a self-review of added code.
- [x] I/we have written code that is legible and maintainable by others.
- [x] I/we have commented the added code, particularly in hard-to-understand areas.
- [x] I/we have made any necessary changes to documentation.
- [x] I/we have added tests that cover new code.
- [x] I/we have run tests and they pass locally with the changes.
- [x] I/we have run `go fmt ./...` and `golangci-lint run`.